Ignore godoc examples by default
godoc examples (https://blog.golang.org/examples) depend on being able to write
to stdout, which is in contradiction to the default forbidden pattern of fmt.Print*.
To make this work nicely out of the box, we will skip checking godoc examples, by default.
This includes "whole file examples", in which an entire file is ignored because
it has only a single Example* method.

This PR also allows the `//permit` directive to be ignored, allowing this
linter to be controlled just using the standard `//nolint` directive from
golangci lint.

This PR also adds a dependency on the launchdarkly/go-options option code
generator, which isn't pinned in the repo.  I may consider doing that at a
later date.

Fixes #3