support import aliases, methods and fields
It was possible to use fmt.Print by importing "fmt" under a different name. Now
a pattern can get matched against the full package path instead of the import
name. This is also useful to ban items from packages that may get imported
under various different names.

It was possible to forbid a function in a package, but not a method or field of
a certain type in a package. The same approach for replacing the match text
can now also be used for those.

The only way of debugging was to single-step with a debugger and manually
printing variables. This gets tiresome because in particular for the analyzer
test, inner functions get called a lot. Debug log output during unit testing
helps. Informing the user about text expansion may also be useful, but is not
addressed yet.