support import aliases and methods
It was possible to use fmt.Print by importing "fmt" under a different name. Now
a pattern can get matched against the full package path instead of the import
name. This is also useful to ban items from packages that may get imported
under various different names.

It was possible to forbid a function in a package, but not a method of a
certain type in a package. The same expanded rule for matching package
selectors also works for variable selectors.