pkgTextFor: simplify implementation
`Types` is actually a `map[ast.Expr]TypeAndValue`, which means that we can
index it with the selector expressions. This works for
`myCustomStruct{}.AlsoForbidden()` (was not handled before) and for function
calls (handled with custom code, which now can be removed). Simple identifiers
still need to be looked up separately.